requirements.txt
torch>=1.7.0
numpy==1.24.3
torchvision
torchaudio
opencv-python
Pillow
pandas
requests
pathlib
transformers
diffusers
accelerate
lama-cleaner
ipykernel
